Add files via upload
The purpose of this work was to develop a problem-solving approach and a simulation
tool that is useful for the specification of wastewater treatment process equipment design parameters.
The proposition of using an artificial neural network (ANN) numerical model for supervised learning
of a dataset and then for process simulation on a new dataset was investigated. The effectiveness
of the approach was assessed by evaluating the capacity of the model to distinguish differences
in the equipment design parameters. To demonstrate the approach, a mock dataset was derived
from experimentally acquired data and physical effects reported in the literature. The mock dataset
comprised the influent flow rate, the bed packing material dimension, the type of packing material
and the packed bed height-to-diameter ratio as predictors of the calorific value reduction. The
multilayer perceptron (MLP) ANN was compared to a polynomial model. The validation test results
show that the MLP model has four hidden layers, each having 256 units (nodes), accurately predicts
calorific value reduction. When the model was fed previously unseen test data, the root-mean-square
error (RMSE) of the predicted responses was 0.101 and the coefficient of determination (R2) was
0.66. The results of simulation of all 125 possible combinations of the 3 mechanical parameters and
identical influent wastewater flow profiles were ranked according to total calorific value reduction.
A t-test of the difference between the mean calorific value reduction of the two highest ranked
experiments showed that the means are significantly different (p-value = 0.011). Thus, the model
has the capacity to distinguish differences in the equipment design parameters. Consequently, the
values of the three mechanical feature parameters from the highest ranked simulated experiment
are recommended for use in the design of the industrial scale upflow anaerobic filter (UAF) for
wastewater treatment.